Humanizada cifra de Población en el detalle de un pais

+++ app/src/main/java/com/salesianostriana/worldquizapp/ui/country/MyCountryRecyclerViewAdapter.java
@@ -95,7 +95,7 @@ public class MyCountryRecyclerViewAdapter extends RecyclerView.Adapter<MyCountry
                 if (null != mListener) {
                     Intent intent = new Intent(ctx, DetailsActivity.class);
                     intent.putExtra("nameCountry", holder.mItem.getName());
-                    intent.putExtra("peopleCountry", holder.mItem.getPopulation());
+                    intent.putExtra("peopleCountry", (double)(holder.mItem.getPopulation()));
                     intent.putExtra("capitalCountry", holder.mItem.getCapital());
                     intent.putExtra("currencyCountry", holder.mItem.getCurrencies().get(0).getCode());
                     intent.putExtra("languageCountry", holder.mItem.getLanguages().get(0).getName());
